Move pure BingTile functions to BingTileUtils
This will keep BingTileFunctions focused on the interaction between
Presto functions and BingTiles.
