+++ presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/BingTileFunctions.java
@@ -26,7 +26,6 @@ import com.facebook.presto.spi.function.Description;
 import com.facebook.presto.spi.function.ScalarFunction;
 import com.facebook.presto.spi.function.ScalarOperator;
 import com.facebook.presto.spi.function.SqlType;
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import io.airlift.slice.Slice;
 
@@ -41,9 +40,30 @@ import static com.facebook.presto.geospatial.GeometryUtils.isPointOrRectangle;
 import static com.facebook.presto.geospatial.serde.EsriGeometrySerde.deserialize;
 import static com.facebook.presto.geospatial.serde.EsriGeometrySerde.serialize;
 import static com.facebook.presto.plugin.geospatial.BingTile.MAX_ZOOM_LEVEL;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.EARTH_RADIUS_KM;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.LATITUDE_OUT_OF_RANGE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.LATITUDE_SPAN_OUT_OF_RANGE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.LONGITUDE_OUT_OF_RANGE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.LONGITUDE_SPAN_OUT_OF_RANGE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.MAX_LATITUDE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.MAX_LONGITUDE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.MIN_LATITUDE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.MIN_LONGITUDE;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.TILE_PIXELS;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.checkCondition;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.checkCoordinate;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.checkLatitude;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.checkLongitude;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.checkQuadKey;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.checkZoomLevel;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.latitudeLongitudeToTile;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.longitudeToTileX;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.longitudeToTileY;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.mapSize;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.tileToEnvelope;
+import static com.facebook.presto.plugin.geospatial.BingTileUtils.tileXYToLatitudeLongitude;
 import static com.facebook.presto.plugin.geospatial.GeometryType.GEOMETRY_TYPE_NAME;
 import static com.facebook.presto.spi.StandardErrorCode.INVALID_CAST_ARGUMENT;
-import static com.facebook.presto.spi.StandardErrorCode.INVALID_FUNCTION_ARGUMENT;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Verify.verify;
 import static io.airlift.slice.Slices.utf8Slice;
@@ -66,25 +86,8 @@ import static java.lang.String.format;
  */
 public class BingTileFunctions
 {
-    private static final int TILE_PIXELS = 256;
-
-    @VisibleForTesting
-    static final double MAX_LATITUDE = 85.05112878;
-    private static final double MIN_LATITUDE = -85.05112878;
-    @VisibleForTesting
-    static final double MIN_LONGITUDE = -180;
-    private static final double MAX_LONGITUDE = 180;
-    private static final double EARTH_RADIUS_KM = 6371.01;
     private static final int OPTIMIZED_TILING_MIN_ZOOM_LEVEL = 10;
     private static final Block EMPTY_TILE_ARRAY = BIGINT.createFixedSizeBlockBuilder(0).build();
-
-    private static final String LATITUDE_OUT_OF_RANGE = "Latitude must be between " + MIN_LATITUDE + " and " + MAX_LATITUDE;
-    private static final String LATITUDE_SPAN_OUT_OF_RANGE = String.format("Latitude span for the geometry must be in [%.2f, %.2f] range", MIN_LATITUDE, MAX_LATITUDE);
-    private static final String LONGITUDE_OUT_OF_RANGE = "Longitude must be between " + MIN_LONGITUDE + " and " + MAX_LONGITUDE;
-    private static final String LONGITUDE_SPAN_OUT_OF_RANGE = String.format("Longitude span for the geometry must be in [%.2f, %.2f] range", MIN_LONGITUDE, MAX_LONGITUDE);
-    private static final String QUAD_KEY_TOO_LONG = "QuadKey must be " + MAX_ZOOM_LEVEL + " characters or less";
-    private static final String ZOOM_LEVEL_TOO_SMALL = "Zoom level must be >= 0";
-    private static final String ZOOM_LEVEL_TOO_LARGE = "Zoom level must be <= " + MAX_ZOOM_LEVEL;
 
     private BingTileFunctions() {}
 
@@ -598,97 +601,6 @@ public class BingTileFunctions
         }
     }
 
-    private static Point tileXYToLatitudeLongitude(int tileX, int tileY, int zoomLevel)
-    {
-        long mapSize = mapSize(zoomLevel);
-        double x = (clip(tileX * TILE_PIXELS, 0, mapSize) / mapSize) - 0.5;
-        double y = 0.5 - (clip(tileY * TILE_PIXELS, 0, mapSize) / mapSize);
-
-        double latitude = 90 - 360 * Math.atan(Math.exp(-y * 2 * Math.PI)) / Math.PI;
-        double longitude = 360 * x;
-        return new Point(longitude, latitude);
-    }
-
-    /**
-     * Returns a Bing tile at a given zoom level containing a point at a given latitude and longitude.
-     * Latitude must be within [-85.05112878, 85.05112878] range. Longitude must be within [-180, 180] range.
-     * Zoom levels from 1 to 23 are supported.
-     */
-    private static BingTile latitudeLongitudeToTile(double latitude, double longitude, int zoomLevel)
-    {
-        long mapSize = mapSize(zoomLevel);
-        int tileX = longitudeToTileX(longitude, mapSize);
-        int tileY = longitudeToTileY(latitude, mapSize);
-        return BingTile.fromCoordinates(tileX, tileY, zoomLevel);
-    }
-
-    /**
-     * Given latitude and longitude in degrees, and the level of detail, the pixel XY coordinates can be calculated as follows:
-     * sinLatitude = sin(latitude * pi/180)
-     * pixelX = ((longitude + 180) / 360) * 256 * 2level
-     * pixelY = (0.5 – log((1 + sinLatitude) / (1 – sinLatitude)) / (4 * pi)) * 256 * 2level
-     * The latitude and longitude are assumed to be on the WGS 84 datum. Even though Bing Maps uses a spherical projection,
-     * it’s important to convert all geographic coordinates into a common datum, and WGS 84 was chosen to be that datum.
-     * The longitude is assumed to range from -180 to +180 degrees, and the latitude must be clipped to range from -85.05112878 to 85.05112878.
-     * This avoids a singularity at the poles, and it causes the projected map to be square.
-     * <p>
-     * reference: https://msdn.microsoft.com/en-us/library/bb259689.aspx
-     */
-    private static int longitudeToTileX(double longitude, long mapSize)
-    {
-        double x = (longitude + 180) / 360;
-        return axisToCoordinates(x, mapSize);
-    }
-
-    private static int longitudeToTileY(double latitude, long mapSize)
-    {
-        double sinLatitude = Math.sin(latitude * Math.PI / 180);
-        double y = 0.5 - Math.log((1 + sinLatitude) / (1 - sinLatitude)) / (4 * Math.PI);
-        return axisToCoordinates(y, mapSize);
-    }
-
-    /**
-     * Take axis and convert it to Tile coordinates
-     */
-    private static int axisToCoordinates(double axis, long mapSize)
-    {
-        int tileAxis = (int) clip(axis * mapSize, 0, mapSize - 1);
-        return tileAxis / TILE_PIXELS;
-    }
-
-    private static Envelope tileToEnvelope(BingTile tile)
-    {
-        Point upperLeftCorner = tileXYToLatitudeLongitude(tile.getX(), tile.getY(), tile.getZoomLevel());
-        Point lowerRightCorner = tileXYToLatitudeLongitude(tile.getX() + 1, tile.getY() + 1, tile.getZoomLevel());
-        return new Envelope(upperLeftCorner.getX(), lowerRightCorner.getY(), lowerRightCorner.getX(), upperLeftCorner.getY());
-    }
-
-    private static void checkZoomLevel(long zoomLevel)
-    {
-        checkCondition(zoomLevel >= 0, ZOOM_LEVEL_TOO_SMALL);
-        checkCondition(zoomLevel <= MAX_ZOOM_LEVEL, ZOOM_LEVEL_TOO_LARGE);
-    }
-
-    private static void checkCoordinate(long coordinate, long zoomLevel)
-    {
-        checkCondition(coordinate >= 0 && coordinate < (1 << zoomLevel), "XY coordinates for a Bing tile at zoom level %s must be within [0, %s) range", zoomLevel, 1 << zoomLevel);
-    }
-
-    private static void checkQuadKey(@SqlType(StandardTypes.VARCHAR) Slice quadkey)
-    {
-        checkCondition(quadkey.length() <= MAX_ZOOM_LEVEL, QUAD_KEY_TOO_LONG);
-    }
-
-    private static void checkLatitude(double latitude, String errorMessage)
-    {
-        checkCondition(latitude >= MIN_LATITUDE && latitude <= MAX_LATITUDE, errorMessage);
-    }
-
-    private static void checkLongitude(double longitude, String errorMessage)
-    {
-        checkCondition(longitude >= MIN_LONGITUDE && longitude <= MAX_LONGITUDE, errorMessage);
-    }
-
     private static boolean withinDistance(GreatCircleDistanceToPoint distanceFunction, double maxDistance, Point point)
     {
         return distanceFunction.distance(point.getY(), point.getX()) <= maxDistance;
@@ -724,22 +636,5 @@ public class BingTileFunctions
             double t3 = sinLatitude * sin2 + cosLatitude * cos2 * cosDeltaLongitude;
             return atan2(sqrt(t1 * t1 + t2 * t2), t3) * EARTH_RADIUS_KM;
         }
-    }
-
-    private static void checkCondition(boolean condition, String formatString, Object... args)
-    {
-        if (!condition) {
-            throw new PrestoException(INVALID_FUNCTION_ARGUMENT, format(formatString, args));
-        }
-    }
-
-    private static double clip(double n, double minValue, double maxValue)
-    {
-        return Math.min(Math.max(n, minValue), maxValue);
-    }
-
-    private static long mapSize(int zoomLevel)
-    {
-        return 256L << zoomLevel;
     }
 }
